MIX-263 refactor: split tracklist game.tsx into feature sub-components
